+++ src/components/UploadWallPaper/UploadWallPaperHeader.scss
@@ -1,6 +1,6 @@
 @import '../../styles/utils.scss';
 
-.uploadHeader-wrapper{
+.uploadWallPaperHeader-wrapper{
     width:100%;
     padding:0 1.6rem;
     background-color: white;
