+++ src/components/UploadWallPaper/UploadWallPaperQuestion.scss
@@ -1,6 +1,6 @@
 @import '../../styles/utils.scss';
 
-.uploadQuestion-wrapper{
+.uploadWallPaperQuestion-wrapper{
     width:100%;
     display: flex;
     align-items: center;
