+++ src/components/UploadWallPaper/UploadWallPaperToggle.scss
@@ -1,0 +1,1 @@
+@import '../../styles/utils.scss';
