+++ src/components/UploadWallPaper/UploadWallPaperToggle.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import './UploadWallPaperToggle.scss';
+
+function UploadWallPaperToggle() {
+  return <div>hello</div>;
+}
+
+export default UploadWallPaperToggle;
