+++ src/pages/App.tsx
@@ -7,7 +7,7 @@ import IntroPage from './IntroPage';
 import ArchivingPage from './ArchivingPage';
 import Home from './Home/Home';
 import MyPage from './MyPage/MyPage';
-import Upload from './UploadPage/Upload';
+import UploadWallPaper from './UploadWallPaper/UploadWallPaper';
 
 function App() {
   return (
@@ -20,7 +20,7 @@ function App() {
         <Route path="/survey" element={<Survey />} />
         <Route path="/home" element={<Home />} />
         <Route path="/mypage" element={<MyPage />} />
-        <Route path="/upload" element={<Upload />} />
+        <Route path="/upload-wallpaper" element={<UploadWallPaper />} />
       </Routes>
     </Router>
   );
