+++ src/pages/UploadWallPaper/UploadWallPaper.scss
@@ -1,6 +1,6 @@
 @import '../../styles/utils.scss';
 
-.upload-wrapper{
+.uploadWallPaper-wrapper{
 
     main{
         padding:5.8rem 1.6rem 1.6rem;
